Revise and export remove_adjacent_trips
This changes remove_adjacent_trips and elevates it to a user-exposed function. This can be handy when documenting the removal process to clients step-by-step. The result is now also a list containing some diagnostic information, and the centroids input can be replaced with an externals flag on the districts layer.